Fix:Se cambiaron los nombres de los permisos del sistema como tambien se aumentaron los actuales

+++ resources/views/administrador/usuarios/roles.blade.php
@@ -11,9 +11,11 @@
                             <h4 class="card-title">LISTA DE ROLES</h4>
                         </div>
                         <div class="col-auto">
-                            <button class="btn btn-primary" onclick="abrirModalRol()">
-                                <i class="fas fa-plus me-1"></i> Nuevo
-                            </button>
+                            @can('admin.rol.crear')
+                                <button class="btn btn-primary" onclick="abrirModalRol()">
+                                    <i class="fas fa-plus me-1"></i> Nuevo
+                                </button>
+                            @endcan
                         </div>
                     </div>
                 </div>
@@ -41,21 +43,29 @@
                                 class="rounded-circle thumb-xxl">
                         </div>
                         <div class="mt-3">
-                            <a href="#" class="btn btn-sm btn-outline-warning px-2 d-inline-flex align-items-center"
-                                onclick="abrirModalRol('{{ $lis->id }}')">
-                                <i class="iconoir-warning-circle fs-14 me-1"></i>
-                                Editar
-                            </a>
-                            <a href="#" class="btn btn-sm btn-outline-danger px-2 d-inline-flex align-items-center"
-                                onclick="eliminarRol('{{ $lis->id }}')">
-                                <i class="iconoir-trash fs-14 me-1"></i>
-                                Eliminar
-                            </a>
-                            <a href="#" class="btn btn-sm btn-outline-secondary px-2 d-inline-flex align-items-center"
-                                onclick="vizualizarRolPermiso('{{ $lis->id }}')">
-                                <i class="iconoir-eye fs-14 me-1"></i>
-                                Vizualizar
-                            </a>
+                            @can('admin.rol.editar')
+                                <a href="#" class="btn btn-sm btn-outline-warning px-2 d-inline-flex align-items-center"
+                                    onclick="abrirModalRol('{{ $lis->id }}')">
+                                    <i class="iconoir-warning-circle fs-14 me-1"></i>
+                                    Editar
+                                </a>
+                            @endcan
+
+                            @can('admin.rol.eliminar')
+                                <a href="#" class="btn btn-sm btn-outline-danger px-2 d-inline-flex align-items-center"
+                                    onclick="eliminarRol('{{ $lis->id }}')">
+                                    <i class="iconoir-trash fs-14 me-1"></i>
+                                    Eliminar
+                                </a>
+                            @endcan
+
+                            @can('admin.rol.visualizar')
+                                <a href="#" class="btn btn-sm btn-outline-secondary px-2 d-inline-flex align-items-center"
+                                    onclick="vizualizarRolPermiso('{{ $lis->id }}')">
+                                    <i class="iconoir-eye fs-14 me-1"></i>
+                                    Vizualizar
+                                </a>
+                            @endcan
                         </div>
                     </div>
                 </div>
@@ -310,8 +320,8 @@
             })
         }
 
-        let rolNombre       = document.getElementById('rolNombre');
-        let permisoListar   = document.getElementById('permisoListar');
+        let rolNombre = document.getElementById('rolNombre');
+        let permisoListar = document.getElementById('permisoListar');
         async function vizualizarRolPermiso(id) {
             permisoListar.innerHTML = '';
             try {
@@ -324,7 +334,7 @@
                 });
 
                 let data = await response.json();
-                if(data.tipo === 'success'){
+                if (data.tipo === 'success') {
                     modal_vizualizar.show();
                     rolNombre.innerHTML = data.mensaje.rol;
                     data.mensaje.permisos.forEach(elem => {
@@ -333,7 +343,7 @@
                         `;
                     });
                 }
-                if(data.tipo === 'error'){
+                if (data.tipo === 'error') {
                     alerta_top(data.tipo, data.mensaje);
                 }
